Transition Funtionality Change
Incorporated the capturing of the hash in the URL in order to help to detect when the user clicks an A link that leads to the same. Then if they do, instead of using the transition, the page will smooth scroll there. Then the transition is used if a new page is actually loaded